optimize event dashboard

+++ client/src/pages/admin/EventDashboard/EventList/index.jsx
@@ -1,7 +1,33 @@
 import { useDispatch, useSelector } from "react-redux";
-import { deleteEvent, getEvents } from "../eventSlice";
+import { getEvents, deleteEvent } from "../eventSlice";
+import { eventAPI } from "api";
+import { deleteObject, ref } from "firebase/storage";
+import { storage } from "config/firebase";
 import style from "./EventList.module.scss";
 import Button from "components/Button";
+
+const Status = ({ start, end }) => {
+  const now = new Date().toISOString();
+  if (now < start) {
+    return (
+      <span className={style.upcoming}>
+        <i className="bi bi-circle-fill"></i> Upcoming
+      </span>
+    );
+  } else if (now < end) {
+    return (
+      <span className={style.ongoing}>
+        <i className="bi bi-circle-fill"></i> Ongoing
+      </span>
+    );
+  } else {
+    return (
+      <span className={style.ended}>
+        <i className="bi bi-circle-fill"></i> Ended
+      </span>
+    );
+  }
+};
 
 const EventList = ({ setId }) => {
   const dispatch = useDispatch();
@@ -11,29 +37,15 @@ const EventList = ({ setId }) => {
     dispatch(getEvents({ skip: events.length, limit: 5 }));
   };
   const deleteHandle = (id) => {
-    dispatch(deleteEvent(id));
-  };
-
-  const Status = ({ start, end }) => {
-    const now = new Date().toISOString();
-    if (now < start) {
-      return (
-        <span className={style.upcoming}>
-          <i className="bi bi-circle-fill"></i> Upcoming
-        </span>
-      );
-    } else if (now < end) {
-      return (
-        <span className={style.ongoing}>
-          <i className="bi bi-circle-fill"></i> Ongoing
-        </span>
-      );
-    } else {
-      return (
-        <span className={style.ended}>
-          <i className="bi bi-circle-fill"></i> Ended
-        </span>
-      );
+    if (window.confirm("Are you sure to delete this event")) {
+      eventAPI.getEvent(id).then((res) => {
+        res.data.images &&
+          res.data.images.forEach((image) => {
+            const currentRef = ref(storage, image);
+            deleteObject(currentRef);
+          });
+        dispatch(deleteEvent(id));
+      });
     }
   };
